+++ tests/.prettierrc.json
@@ -1,6 +1,5 @@
 {
   "singleQuote": true,
-  "maxLength": 80,
   "semi": true,
   "tabWidth": 2
 }
